Ex 4: state management with a store
First of all, we need to install the Vuex package. To do this, we need to run the following command in the terminal:

```bash
npm install vuex@next
```

Vuex has been added to the project. We created a store at `src/store/index.ts`, added the `user` state to the store, and imported at `main.ts`.
Props and event related to `user` has been removed in all components and replaced by the store when needed

Here's the result:
![Logged as](https://raw.githubusercontent.com/LucasKoc/TP4-ST2AWD/refs/heads/main/Ressources/Screenshot%202024-10-15%20at%2022.36.55.png "Logged as user from store")